+++ src/evm-block-extractor/tests/client/mod.rs
@@ -1,4 +1,6 @@
-use gcp_bigquery_client::Client;
+use std::sync::Arc;
+
+use evm_block_extractor::database::{DatabaseClient, big_query_db_client::BigQueryDbClient};
 use tempfile::NamedTempFile;
 use testcontainers::testcontainers::{clients::Cli, core::WaitFor, Container, GenericImage};
 
@@ -6,13 +8,15 @@ use self::auth_mock::GoogleAuthMock;
 
 pub async fn new_bigquery_client<'a>(
     docker: &'a Cli,
-    project_id: &str,
 ) -> (
-    Client,
+    Arc<dyn DatabaseClient>,
     Container<'a, GenericImage>,
     NamedTempFile,
     GoogleAuthMock,
 ) {
+    let project_id = format!("test_project_{}", rand::random::<u64>());
+    let dataset_id = format!("test_{}", rand::random::<u64>());
+
     let emulator_image = GenericImage::new("ghcr.io/goccy/bigquery-emulator", "0.4.4")
         .with_exposed_port(9050)
         .with_wait_for(WaitFor::message_on_stdout(
@@ -41,7 +45,16 @@ pub async fn new_bigquery_client<'a>(
         .await
         .unwrap();
 
-    (gcp_client, node, temp_file, google_auth)
+        let client = Arc::new(
+            BigQueryDbClient::new_with_client(
+                project_id.clone(),
+                dataset_id.clone(),
+                gcp_client,
+            )
+            .unwrap(),
+        );
+
+    (client, node, temp_file, google_auth)
 }
 
 pub mod auth_mock {
