+++ src/evm-block-extractor/tests/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod block_extractor_it;
+pub mod database_client_it;
